requirements.txt
Pillow
gunicorn
django-heroku
python-dotenv
requests
pyjwt
selenium
webdriver-manager